% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_my_app.R
\name{run_my_app}
\alias{run_my_app}
\title{Launch the Shiny app}
\usage{
run_my_app()
}
\value{
No return value. This function is called for its side effect
of launching a Shiny app.
}
\description{
Opens the interactive temperature anomaly explorer.
The app lets you:
\itemize{
\item Filter by country (e.g. World, Australia)
\item Choose a year range
\item View a time series plot of temperature anomaly
\item Inspect the underlying data table
}
}
